requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
temporalio>=1.7.0
requests>=2.31.0
pydantic>=2.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
strands-agents
ollama
